refactor(copyTemplateFiles): move copyTemplateFiles into separate file
